fix(migration): address PR #169 review comments
Backend:
- Use authenticated principal (ignore user_id query param) in /trips index
- Omit member email for public visitors in full-trip serializer
- Resolve polymorphic comment targets (title/name) in serializer
- Add boolean casts to Trip section-visibility flags
- Add missing activityDragEnd/activityDuplicate/by-id handlers
- Trip-scoped + by-id content delete now cleans comment graphs
- Comment store rejects cross-trip group IDs; direct delete cleans empty group
- Task create via trip-less route; updateById persists explicit nulls; widen index columns
- Member update resolves pivot row by id; remove-by-id fixed
- Sync: add trip_user/comment_group_objects/comments sources + composite cursor
- Import: verify-config fails on missing expected entities
- Remove stale npm from composer setup; fix README example URLs
- Enforce CSRF on auth routes; frontend auth uses CSRF mutations

Frontend:
- Gate Trips/TripsPublic reads behind flag with InstantDB fallback
- Swap-day uses PATCH; logout uses POST; account upgrade uses CSRF mutation
- apiTrip coerces MySQL string numerics; gated trip read stores
- Update UserFactory to migrated user schema

+++ backend/database/migrations/2026_01_01_000003_create_comments_and_tasks_tables.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->string('id', 40)->primary();
             $table->string('trip_id', 40);
             $table->string('title');
-            $table->tinyInteger('index');
+            $table->integer('index');
             $table->tinyInteger('status'); // open / done
             $table->bigInteger('created_at_ms');
             $table->bigInteger('updated_at_ms');
@@ -22,7 +22,7 @@ return new class extends Migration
         Schema::create('tasks', function (Blueprint $table): void {
             $table->string('id', 40)->primary();
             $table->string('task_list_id', 40);
-            $table->tinyInteger('index');
+            $table->integer('index');
             $table->string('title');
             $table->text('description')->nullable();
             $table->tinyInteger('status'); // 0 todo / 1 done
